+++ lib/tests/test_revset_graph_iterator.rs
@@ -13,8 +13,8 @@
 // limitations under the License.
 
 use itertools::Itertools;
+use jujutsu_lib::default_revset_engine::revset_for_commits;
 use jujutsu_lib::repo::Repo;
-use jujutsu_lib::revset::revset_for_commits;
 use jujutsu_lib::revset_graph_iterator::{RevsetGraphEdge, RevsetGraphIterator};
 use test_case::test_case;
 use testutils::{CommitGraphBuilder, TestRepo};
